refactoring related with `bittensor.utils.weight_utils.convert_and_normalize_weights_and_uids`
